Title: 나눗셈의 나머지, Time: 114ms, Memory: 24MB, Status: Passed - CodeTree

+++ 241014/나눗셈의 나머지/remainder-of-division.py
@@ -3,7 +3,7 @@ cnt = []
 count = 0
 while a > 1 :
     cnt.append(int(a%b))
-    a = a/b
+    a = int(a/b)
 
 for i in range(b) :
     count += cnt.count(i)**2
